// model/Currency.ts
export enum Currency {
    BTC = 'BTC',
    BCH = 'BCH',
    LTC = 'LTC',
    ETH = 'ETH',
    XRP = 'XRP',
    XLM = 'XLM',
    VET = 'VET',
    NEO = 'NEO',
    BNB = 'BNB',
    USDT = 'USDT',
    LEO = 'LEO',
    LINK = 'LINK',
    FREE = 'FREE',
    MKR = 'MKR',
    USDC = 'USDC',
    BAT = 'BAT',
    TUSD = 'TUSD',
    PAX = 'PAX',
    PLTC = 'PLTC',
    XCON = 'XCON',
    MMY = 'MMY',
    PAXG = 'PAXG',
}

export const ETH_BASED_CURRENCIES = [
    Currency.USDT.toString(),
    Currency.LEO.toString(),
    Currency.LINK.toString(),
    Currency.FREE.toString(),
    Currency.MKR.toString(),
    Currency.USDC.toString(),
    Currency.BAT.toString(),
    Currency.TUSD.toString(),
    Currency.PAX.toString(),
    Currency.PAXG.toString(),
    Currency.PLTC.toString(),
    Currency.XCON.toString(),
    Currency.ETH.toString(),
    Currency.MMY.toString(),
];